+++ 2.Logic-Layer/PresupuestosServices/src/vo/vista/UserListJSON.java
@@ -2,7 +2,7 @@ package vo.vista;
 
 public class UserListJSON {
 	private int iduser;
-	private int document;
+	private long document;
 	private String name;
 	private String username;
 	private String roll;
@@ -21,10 +21,10 @@ public class UserListJSON {
 	public void setIduser(int iduser) {
 		this.iduser = iduser;
 	}
-	public int getDocument() {
+	public long getDocument() {
 		return document;
 	}
-	public void setDocument(int document) {
+	public void setDocument(long document) {
 		this.document = document;
 	}
 	public String getName() {
